Fix for delete with warning (missing confirmation before api call)

+++ app/js/app.js
@@ -4978,6 +4978,7 @@ function runBlock() {
         vm.saveSection = _saveSection;
         vm.editSection = _editSection;
         vm.removeSection = _removeSection;
+        vm.cancelSection = _cancelSection;
 
         initialize();
 
@@ -5182,20 +5183,25 @@ function runBlock() {
             vm.selected_section.form = vm.formId;
             vm.showSectionForm = true;
         }
+        function _cancelSection() {
+            vm.showSectionForm = false;
+        }
 
         function _removeSection(section) {
-            AlertService.showConfirmForDelete("You are about to DELETE SECTION",
-                "Are you sure?", "Yes, Delete it!", "warning", true,function () {
+            AlertService.showConfirmForDelete("You are about to DELETE SECTION, All Questions under this section will be removed",
+                "Are you sure?", "Yes, Delete it!", "warning", true,function (isConfirm) {
+                    if(isConfirm){
+                        FormService.RemoveSection(section).then(function(response){
+                            vm.showSectionForm = false;
+                            callAPI();
+                            AlertService.showSuccess("SECTION","Section Deleted successfully");
+                        },function(error){
+                            console.log("qn deleting error",error);
+                            var message = error.data.error.message;
+                            AlertService.showError("Failed to DELETE Section",message);
+                        });
+                    }
 
-                    FormService.RemoveSection(section).then(function(response){
-                        vm.showSectionForm = false;
-                        callAPI();
-                        AlertService.showSuccess("SECTION","Section Deleted successfully");
-                    },function(error){
-                        console.log("qn deleting error",error);
-                        var message = error.data.error.message;
-                        AlertService.showError("Failed to DELETE Section",message);
-                    });
                 });
         }
     }
@@ -5331,7 +5337,9 @@ function runBlock() {
         }
         function _removeQuestion() {
             AlertService.showConfirmForDelete("You are about to DELETE this Question?",
-                "Are you sure?", "Yes, Delete it!", "warning", true,function () {
+                "Are you sure?", "Yes, Delete it!", "warning", true,function (isConfirm) {
+
+                if(isConfirm){
                     FormService.DeleteQuestion(vm.question).then(function(response){
                         AlertService.showSuccess("Question","Question Deleted successfully");
                         $mdDialog.hide();
@@ -5340,6 +5348,8 @@ function runBlock() {
                         var message = error.data.error.message;
                         AlertService.showError("Failed to DELETE Question",message);
                     })
+                }
+
                 });
 
         }
